Split MCP client responsibilities
